DB Modified, added a Delete.html, modified the views.py to render the get and post methods for this Delete.html form.

+++ OpenHouse/base/views.py
@@ -59,6 +59,10 @@ def updateRoom(request, pk):
 
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
-    room.delete()
-    return redirect('home')
+    if request.method=="POST":
+        room.delete()
+        return redirect('home')
+    context = {'obj':room}
+    return render(request,'base/Delete.html', context)
+
